+++ CHANGELOG.md
@@ -2,7 +2,8 @@
  * Added Regex, Inst, and Uuid as primitive schema types (thanks [jwhitlark](https://github.com/jwhitlark))
  * Add annotated arglists to functions defined with `s/defn` (thanks [danielneal](https://github.com/danielneal))
  * Add `set-fn-validation!` to schema.core, to globally turn validation on or off.
-
+ * Add `:always-validate` metadata on fn/defn name to unconditionally use validation.
+	
 ## 0.1.3
  * Fix compatibility with Clojurescript 1889 (removal of format)